Rename field, move class to own file
