+++ trunk/source/BuchiAutomizer/src/de/uni_freiburg/informatik/ultimate/plugins/generator/buchiautomizer/BuchiCegarLoopBenchmark.java
@@ -31,9 +31,7 @@ import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.HashSet;
-import java.util.LinkedHashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.Set;
 
 import de.uni_freiburg.informatik.ultimate.lassoranker.LassoAnalysis.PreprocessingBenchmark;
@@ -47,7 +45,6 @@ import de.uni_freiburg.informatik.ultimate.plugins.generator.traceabstraction.Ce
 import de.uni_freiburg.informatik.ultimate.util.csv.CsvUtils;
 import de.uni_freiburg.informatik.ultimate.util.csv.ICsvProvider;
 import de.uni_freiburg.informatik.ultimate.util.statistics.IStatisticsDataProvider;
-import de.uni_freiburg.informatik.ultimate.util.statistics.IStatisticsType;
 
 public class BuchiCegarLoopBenchmark extends CegarStatisticsType {
 
@@ -369,128 +366,4 @@ public class BuchiCegarLoopBenchmark extends CegarStatisticsType {
 		}
 		return mostDifficult;
 	}
-
-	public static class LassoAnalysisResults implements IStatisticsDataProvider, IStatisticsType {
-		public static final String LASSO_NONTERMINATING = "nont";
-		public static final String TERMINATION_UNKNOWN = "unkn";
-		/**
-		 * Cases where (already a single iteration of) the loop is infeasible.
-		 */
-		public static final String STEM_FEASIBLE_LOOP_INFEASIBLE = "SFLI";
-		/**
-		 * Cases where the stem is feasible, (a single iteration of) the loop is feasible but the loop is terminating.
-		 */
-		public static final String STEM_FEASIBLE_LOOP_TERMINATING = "SFLT";
-		/**
-		 * Cases where stem and loop are feasible but the concatenation of stem and loop is infeasible.
-		 */
-		public static final String CONCATENATION_INFEASIBLE = "conc";
-		/**
-		 * Cases where stem and loop are feasible but the concatenation of stem and loop is infeasible and the loop is
-		 * terminating.
-		 */
-		public static final String CONCATENATION_INFEASIBLE_LOOP_TERMINATING = "concLT";
-		/**
-		 * Cases where the stem is infeasible and the loop is nonterminating.
-		 */
-		public static final String STEM_INFEASIBLE_LOOP_NONTERMINATING = "SILN";
-		/**
-		 * Cases where the stem is infeasible and the termination/feasibility of the loop is unknown.
-		 */
-		public static final String STEM_FEASIBLE_LOOP_UNKNOWN = "SILU";
-		/**
-		 * Cases where the stem is infeasible and the loop is infeasible.
-		 */
-		public static final String STEM_INFEASIBLE_LOOP_INFEASIBLE = "SILI";
-		/**
-		 * Cases where both, stem and loop are infeasible.
-		 */
-		public static final String STEM_INFEASIBLE_LOOP_TERMINATING = "SILT";
-		/**
-		 * Cases where the stem and the loop are feasible, the loop itself is nonterminating but the lasso is
-		 * terminating.
-		 */
-		public static final String s_LassoTerminating = "lasso";
-
-		public final Map<String, Integer> mMap;
-
-		public LassoAnalysisResults() {
-			mMap = new LinkedHashMap<>();
-			mMap.put(LASSO_NONTERMINATING, 0);
-			mMap.put(TERMINATION_UNKNOWN, 0);
-			mMap.put(STEM_FEASIBLE_LOOP_INFEASIBLE, 0);
-			mMap.put(STEM_FEASIBLE_LOOP_TERMINATING, 0);
-			mMap.put(CONCATENATION_INFEASIBLE, 0);
-			mMap.put(CONCATENATION_INFEASIBLE_LOOP_TERMINATING, 0);
-			mMap.put(STEM_INFEASIBLE_LOOP_NONTERMINATING, 0);
-			mMap.put(STEM_FEASIBLE_LOOP_UNKNOWN, 0);
-			mMap.put(STEM_INFEASIBLE_LOOP_INFEASIBLE, 0);
-			mMap.put(STEM_INFEASIBLE_LOOP_TERMINATING, 0);
-			mMap.put(s_LassoTerminating, 0);
-		}
-
-		@Override
-		public String toString() {
-			final StringBuilder sb = new StringBuilder();
-			for (final String key : getKeys()) {
-				sb.append(key);
-				sb.append(getValue(key));
-				sb.append(" ");
-			}
-			return sb.toString();
-		}
-
-		public void increment(final String key) {
-			final int value = mMap.get(key);
-			mMap.put(key, value + 1);
-		}
-
-		// public void aggregate(LassoAnalysisResults lassoAnalysisResults) {
-		// mLassoNonterminating = lassoAnalysisResults.mLassoNonterminating;
-		// mTerminationUnknown = lassoAnalysisResults.mTerminationUnknown;
-		// mStemFeasibleLoopInfeasible = lassoAnalysisResults.mStemFeasibleLoopInfeasible;
-		// mStemFeasibleLoopTerminating = lassoAnalysisResults.mStemFeasibleLoopTerminating;
-		// mConcatenationInfeasible = lassoAnalysisResults.mConcatenationInfeasible;
-		// mConcatInfeasibleLoopTerminating = lassoAnalysisResults.mConcatInfeasibleLoopTerminating;
-		// mStemInfeasibleLoopNonterminating = lassoAnalysisResults.mStemInfeasibleLoopNonterminating;
-		// mStemInfeasibleLoopUnknown = lassoAnalysisResults.mStemInfeasibleLoopUnknown;
-		// mStemInfeasibleLoopInfeasible = lassoAnalysisResults.mStemInfeasibleLoopInfeasible;
-		// mStemInfeasibleLoopTerminating = lassoAnalysisResults.mStemInfeasibleLoopTerminating;
-		// mLassoTerminating = lassoAnalysisResults.mLassoTerminating;
-		// }
-
-		@Override
-		public Collection<String> getKeys() {
-			return mMap.keySet();
-		}
-
-		@Override
-		public Object getValue(final String key) {
-			return mMap.get(key);
-		}
-
-		@Override
-		public IStatisticsType getBenchmarkType() {
-			return this;
-		}
-
-		@Override
-		public Object aggregate(final String key, final Object value1, final Object value2) {
-			throw new AssertionError("not yet implemented");
-		}
-
-		@Override
-		public String prettyprintBenchmarkData(final IStatisticsDataProvider benchmarkData) {
-			final LassoAnalysisResults lar = (LassoAnalysisResults) benchmarkData;
-			final StringBuilder sb = new StringBuilder();
-			for (final String key : lar.getKeys()) {
-				sb.append(key);
-				sb.append(lar.getValue(key));
-				sb.append(" ");
-			}
-			return sb.toString();
-		}
-
-	}
-
 }
